requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==5.0.1
attrs==24.2.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.0
charset-normalizer==3.4.0
cryptography==36.0.1
Django==3.2.9
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==21.1
django-phonenumber-field==6.0.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
django-utils-six==2.0
djangorestframework==3.12.4
firebase-admin==6.6.0
frozenlist==1.5.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
multidict==6.1.0
Pillow==8.4.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
PyJWT==2.10.1
pyparsing==3.2.0
pytz==2021.3
requests==2.32.3
rsa==4.9
sqlparse==0.4.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.18.3
